Move command operations to cli crate
