+++ crates/huak_cli/src/cmd/activate.rs
