+++ crates/huak_cli/src/cmd/add.rs
@@ -1,6 +1,5 @@
-use crate::{
-    dependency::{dependency_iter, Dependency},
-    Config, HuakResult, InstallOptions,
+use huak_ops::{
+    dependency_iter, Config, Dependency, HuakResult, InstallOptions,
 };
 use pep440_rs::VersionSpecifiers;
 use pep508_rs::VersionOrUrl;
@@ -131,10 +130,10 @@ pub fn add_project_optional_dependencies(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{
+    use huak_ops::{
         fs::{self, CopyDirOptions},
-        ops::{test_config, test_venv},
-        test_resources_dir_path, Verbosity,
+        sys::Verbosity,
+        test::{test_config, test_resources_dir_path, test_venv},
     };
     use tempfile::tempdir;
