+++ crates/huak_cli/src/cmd/version.rs
