+++ crates/huak_ops/src/lib.rs
@@ -53,7 +53,6 @@ mod error;
 mod fs;
 mod git;
 mod metadata;
-pub mod ops;
 mod package;
 mod python_environment;
 mod sys;
